Crate ink_abi

Re-exports

pub use ink_abi_derive::HasLayout;

Structs

ContractSpec
Describes a contract.

DeploySpec
Describes the deploy handler of a contract.

EventParamSpec
Describes a pair of parameter name and type.

EventSpec
Describes an event definition.

InkProject

LayoutField
The layout for a particular field of a struct layout.

LayoutKey
A concrete range of keys.

LayoutRange
Direct range of associated storage keys.

LayoutStruct
A struct storage layout.

MessageParamSpec
Describes a pair of parameter name and type.

MessageSpec
Describes a contract message.

ReturnTypeSpec
Describes the return type of a contract message.

Enums

StorageLayout
Either a concrete layout bound or another layout sub-struct.

Traits

HasLayout
Implemented by types that have a storage layout.